requirements.txt
torch==2.2.*
transformers==4.*
paddlepaddle==2.6.2
paddleocr==2.7.0.3
opencv-python-headless>=4.8
pymupdf
pillow
numpy
streamlit
openai
google-generativeai>=0.7.0
